Move CustomLevel's options panel into its own file.
